+++ llvm/include/llvm/Transforms/Scalar/Reg2Mem.h
@@ -13,11 +13,7 @@
 #ifndef LLVM_TRANSFORMS_SCALAR_REG2MEM_H
 #define LLVM_TRANSFORMS_SCALAR_REG2MEM_H
 
-#include "llvm/Analysis/LoopInfo.h"
-#include "llvm/IR/Dominators.h"
 #include "llvm/IR/PassManager.h"
-#include "llvm/InitializePasses.h"
-#include "llvm/Pass.h"
 
 namespace llvm {
 
@@ -25,27 +21,6 @@ class RegToMemPass : public PassInfoMixin<RegToMemPass> {
 public:
   PreservedAnalyses run(Function &F, FunctionAnalysisManager &AM);
 };
-
-class RegToMemWrapperPass : public FunctionPass {
-public:
-  static char ID;
-
-  RegToMemWrapperPass();
-
-  void getAnalysisUsage(AnalysisUsage &AU) const override {
-    AU.setPreservesAll();
-
-    AU.addPreserved<DominatorTreeWrapperPass>();
-    AU.addRequired<DominatorTreeWrapperPass>();
-
-    AU.addPreserved<LoopInfoWrapperPass>();
-    AU.addRequired<LoopInfoWrapperPass>();
-  }
-
-  bool runOnFunction(Function &F) override;
-};
-
-FunctionPass *createRegToMemWrapperPass();
 
 } // end namespace llvm
